fix(terminal): honor remote Git Bash and split preflight IPC seams
